requirements.txt
scrapy
extruct
itemadapter
feedparser
python-dateutil

kafka-python
afkak

confluent-kafka
requests
fastavro
jsonschema
